Deployed auto_run.py as an exe file for ease of use
