+++ AC Bias/README.txt
@@ -1,22 +1,21 @@
 ----Welcome to the AC Bias HIT Persistent Plotter Tool----
 
 Use Instructions
-1) On a WINDOWS file explorer navigate into the following directory
-    - \\noodle\GRETA\rhoover\python\Code\ccdm\AC Bias
-    - "Shift + Right click" and open "A powershell window here"
+1) On a WINDOWS machine run the "auto_run.exe" file.
+    - exe file launches an ssh session into chimchim to start the script "ac_bias_hit_persistent.py"
+    - exe file handles cleanup of launched item on chimchim
 
-2) Execute run command to start the script
-    - python auto_run.py
+2) While running the script opens a chrome session that refreshes every 5 sec (This can be adjusted)
 
-3) Follow terminal instructions
-    - Script will issue a command via ssh
-
-4) While running the script opens a chrome session that refreshes every 5 sec
-
-5) When done using the tool input "ctrl + c" to exit the script.
+3) When done using the tool input "ctrl + c" (inside the generated window) to exit the application.
     - Exiting kills script on ssh side.
+    - Closing the generated window without entering "ctrl + c" will leave the ssh session script running.
+
+
 
 
 Any questions reach out to:
+
 Ryan Hoover (CCDM)
+484-336-4167
 rhoover@ipa.cfa.harvard.edu
